update edit delet

+++ src/database/migrations/20230814180837-sensors.js
@@ -22,7 +22,7 @@ module.exports = {
         defaultValue: false,
       },
       corretion:{
-        type: Sequelize.INTEGER,
+        type: Sequelize.REAL,
         defaultValue: 0,
       },
       created_at: {
